Fixed the bug where hartree energies for dft were changing but those for hf were constant. This was because rsatom was not being updated after combining the dft and hf wavefunctions.
Now, the test for exxweight=1.0 works as expected.

Also, I have turned off density mixing for rsatom because this should
be done consistently with how the wavefunctions at iteration i are being
mixed with those at iteration i-1. Maybe this is a good time to think
about implementing conjugate gradients ...