Now, the self-interaction correction is calculated like so: ehardft - eharhf where ehardft uses the charge densities and eharhf is the expression using the slater integrals in Froese Fischer book.
Still, the total energies calculated using exxweight = 1.0 are different
from those calculated in the HF code, by about 1e-6 to 1e-7 Ry. This
might be coming from the way we calculate the eigenvalues in hysolve --
we first multiply the wavefunctions by the energies and then calculate
the norm of the wavefunctions to get the eigenvalues. This will give an
error if the norm is not calculated perfectly.