Revert "Merge branch 'master' into singlesolve"
This reverts commit a7b100d68d4e8b0ab254c73db9dc261977cc40e7, reversing
changes made to e8b462680ff19a559a2bd58ef4c31a6eef2e3698.

Please don't merge branches master and singlesolve. They are to be
considered separate because we made changes in master which are
inconsistent with what we are trying to acheive in singlesolve. When
updating github, please use this command:

git push origin singlesolve